refactor: rename 'payload' to analysis domain
