+++ src/analysis/extract.ts
