+++ src/analysis/interpret.ts
