+++ src/pipeline/run.ts
@@ -3,6 +3,10 @@ import type { Scanner } from "@nodesecure/scanner";
 import * as vuln from "@nodesecure/vuln";
 
 import {
+  OutcomePayloadFromPipelineChecks,
+  runPayloadInterpreter
+} from "../analysis/index.js";
+import {
   ApiConfig,
   CliConfig,
   defaultExternalConfigOptions,
@@ -11,10 +15,6 @@ import {
 } from "../config/index.js";
 import { consolePrinter } from "../lib/console-printer/index.js";
 import { Maybe } from "../lib/types/index.js";
-import {
-  OutcomePayloadFromPipelineChecks,
-  runPayloadInterpreter
-} from "../payload/index.js";
 import { scannerReporter, runReporting } from "../reporters/index.js";
 
 import { status } from "./status.js";
